+++ apps/commerce-api/src/test/kotlin/com/loopers/fixture/point/PointFixture.kt
@@ -3,16 +3,15 @@ package com.loopers.fixture.point
 import com.loopers.domain.point.PointCommand
 import com.loopers.domain.point.PointEntity
 import com.loopers.interfaces.api.point.PointV1Models
-import java.math.BigDecimal
 
 sealed class PointFixture {
 
     abstract val userId: String
-    abstract val amount: BigDecimal
+    abstract val amount: Long
 
     fun pointCommand(
         userId: String = this.userId,
-        amount: BigDecimal = this.amount,
+        amount: Long = this.amount,
     ): PointCommand.ChargeInput {
         return PointCommand.ChargeInput(
             userId = userId,
@@ -22,7 +21,7 @@ sealed class PointFixture {
 
     fun pointEntity(
         userId: String = this.userId,
-        amount: BigDecimal = this.amount,
+        amount: Long = this.amount,
     ): PointEntity {
         return PointEntity(
             userId = userId,
@@ -31,13 +30,13 @@ sealed class PointFixture {
     }
 
     fun chargeRequest(
-        amount: BigDecimal = this.amount,
+        amount: Long = this.amount,
     ): PointV1Models.Request.Charge {
         return PointV1Models.Request.Charge(amount = amount)
     }
 
     object Normal: PointFixture() {
         override val userId = "user1"
-        override val amount= BigDecimal.valueOf(23)
+        override val amount= 23L
     }
 }
